+++ src/main/java/org/mesdag/advjs/util/AdvJSEvents.java
@@ -2,8 +2,10 @@ package org.mesdag.advjs.util;
 
 import dev.latvian.mods.kubejs.event.EventGroup;
 import dev.latvian.mods.kubejs.event.EventHandler;
+import net.minecraft.advancements.Advancement;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.storage.loot.LootDataManager;
-import org.mesdag.advjs.AdvJS;
+import net.minecraftforge.fml.ModList;
 import org.mesdag.advjs.advancement.AdvConfigureEventJS;
 import org.mesdag.advjs.advancement.AdvLockEventJS;
 import org.mesdag.advjs.integration.betteradvancements.BetterAdvEventJS;
@@ -11,6 +13,8 @@ import org.mesdag.advjs.integration.betteradvancements.BetterAdvModifier;
 import org.mesdag.advjs.integration.revelationary.AdvRevelationEventJS;
 import org.mesdag.advjs.trigger.Trigger;
 import org.mesdag.advjs.trigger.custom.TriggerRegistryEventJS;
+
+import java.util.Map;
 
 public interface AdvJSEvents {
     EventGroup GROUP = EventGroup.of("AdvJSEvents");
@@ -24,9 +28,12 @@ public interface AdvJSEvents {
 
     static void postAdv(LootDataManager lootData) {
         ADVANCEMENT.post(new AdvConfigureEventJS(new Trigger(lootData)));
-        if (AdvJS.betterLoaded()) {
+    }
+
+    static void postBetter(Map<ResourceLocation, Advancement.Builder> map){
+        if (ModList.get().isLoaded("betteradvancements")) {
             BETTER_ADV.post(new BetterAdvEventJS());
-            BetterAdvModifier.MODIFIERS.removeIf(BetterAdvModifier::modify);
+            BetterAdvModifier.MODIFIERS.removeIf(modifier -> modifier.modify(map.get(modifier.getId())));
         }
     }
 }